Move metadata file container into an independent module
