+++ src/components/@Explore/ExploreTab.tsx
@@ -24,9 +24,9 @@ import { DIDSearchType, IDIDSearchResult } from '../../types';
 import { InlineDropdown } from '../components/../@Explore/InlineDropdown';
 import { ListScopesPopover } from '../components/../@Explore/ListScopesPopover';
 import {
-  IMetadataFilter,
-  MetadataFilterItem
-} from '../components/../@Explore/MetadataFilterItem';
+  buildMetadataFilterString,
+  MetadataFilterContainer
+} from '../components/../@Explore/MetadataFilterContainer';
 
 const useStyles = createUseStyles({
   mainContainer: {
@@ -78,40 +78,6 @@ const useStyles = createUseStyles({
     cursor: 'pointer',
     marginLeft: '4px'
   },
-  addMetadataFilterButton: {
-    marginTop: '4px',
-    marginRight: '16px',
-    marginLeft: '16px',
-    width: 'fit-content',
-    fontSize: '9pt',
-    cursor: 'pointer',
-    opacity: 0.5,
-    '&:hover': {
-      opacity: 1
-    }
-  },
-  deleteFiltersButton: {
-    marginTop: '4px',
-    marginRight: '16px',
-    marginLeft: '16px',
-    width: 'fit-content',
-    display: 'flex',
-    alignItems: 'center',
-    fontSize: '9pt',
-    cursor: 'pointer',
-    opacity: 0.5,
-    '&:hover': {
-      opacity: 1
-    }
-  },
-  deleteFiltersIcon: {
-    fontSize: '9pt',
-    cursor: 'pointer',
-    opacity: 0.5,
-    '&:hover': {
-      opacity: 1
-    }
-  },
   loading: {
     padding: '16px'
   },
@@ -144,22 +110,10 @@ const _Explore: React.FunctionComponent = props => {
   const [didExpanded, setDidExpanded] = useState<{ [index: number]: boolean }>(
     {}
   );
-  const [metadataFilters, setMetadataFilters] = React.useState<
-    IMetadataFilter[]
-  >([]);
   const [error, setError] = useState<string>();
   const [loading, setLoading] = useState(false);
   const [searchTrigger, setSearchTrigger] = useState(0);
   const activeInstance = useStoreState(UIStore, s => s.activeInstance);
-
-  const buildMetadataFilterString = () => {
-    return metadataFilters
-      .map((filter, index) => {
-        const logic = index === 0 ? '' : filter.logic === 'And' ? ',' : ';';
-        return `${logic}${filter.key}${filter.operator}${filter.value}`;
-      })
-      .join('');
-  };
 
   const doSearch = () => {
     setSearchTrigger(prev => prev + 1); // Increment the counter to trigger the search
@@ -270,38 +224,6 @@ const _Explore: React.FunctionComponent = props => {
     );
   };
 
-  const handleAddMetadataFilter = () => {
-    setMetadataFilters([
-      ...metadataFilters,
-      { logic: 'And', key: '', operator: '=', value: '' }
-    ]);
-  };
-
-  const handleDeleteMetadataFilter = (indexToDelete: number) => {
-    setMetadataFilters(prev => {
-      const newFilters = prev.filter((_, index) => index !== indexToDelete);
-      if (indexToDelete === 0 && newFilters.length > 0) {
-        newFilters[0] = { ...newFilters[0], logic: 'And' };
-      }
-      return newFilters;
-    });
-  };
-
-  const handleDeleteAllMetadataFilters = () => {
-    setMetadataFilters([]);
-  };
-
-  const handleFilterChange = (
-    index: number,
-    updatedFilter: IMetadataFilter
-  ) => {
-    setMetadataFilters(prev => {
-      const newFilters = [...prev];
-      newFilters[index] = updatedFilter;
-      return newFilters;
-    });
-  };
-
   return (
     <div className={classes.mainContainer}>
       <div className={classes.searchContainer}>
@@ -325,45 +247,9 @@ const _Explore: React.FunctionComponent = props => {
           optionWidth="180px"
         />
       </div>
-      <div>
-        {!metadataFilters.length && (
-          <div
-            className={classes.addMetadataFilterButton}
-            onClick={handleAddMetadataFilter}
-          >
-            + Add metadata filter
-          </div>
-        )}
-        {metadataFilters.map((filter, index) => (
-          <MetadataFilterItem
-            key={index}
-            filter={filter}
-            showBoolOperatorDropdown={index === 0}
-            onDelete={() => handleDeleteMetadataFilter(index)}
-            onChange={updatedFilter => handleFilterChange(index, updatedFilter)}
-            onKeyPress={handleKeyPress}
-          />
-        ))}
-        {!!metadataFilters.length && (
-          <div
-            className={classes.addMetadataFilterButton}
-            onClick={handleAddMetadataFilter}
-          >
-            + Add filter rule
-          </div>
-        )}
-        {!!metadataFilters.length && (
-          <div
-            className={classes.deleteFiltersButton}
-            onClick={handleDeleteAllMetadataFilters}
-          >
-            <i className={`${classes.deleteFiltersIcon} material-icons`}>
-              delete
-            </i>
-            <span style={{ marginLeft: '8px' }}>Delete filter</span>
-          </div>
-        )}
-      </div>
+      <MetadataFilterContainer
+        onKeyPress={handleKeyPress}
+      />
       {loading && (
         <div className={classes.loading}>
           <Spinning className={`${classes.icon} material-icons`}>
